Move checkbox functionality to filters component
